Make entire project use single region with AWS_REGION
