+++ handel/src/account-config/default-account-config.ts
@@ -14,7 +14,6 @@
  * limitations under the License.
  *
  */
-import * as AWS from 'aws-sdk';
 import { AccountConfig } from 'handel-extension-api';
 import * as extensionSupport from 'handel-extension-support';
 import * as winston from 'winston';
@@ -23,11 +22,6 @@ import * as stsCalls from '../aws/sts-calls';
 
 function getSubnetGroupName(vpcId: string): string {
     return `handel-subnet-groups-${vpcId}`;
-}
-
-async function isValidRegion(region: string) {
-    const regions = await ec2Calls.getRegions();
-    return regions.includes(region);
 }
 
 async function getDefaultVpcSubnets(vpcId: string) {
@@ -55,34 +49,25 @@ async function getSubnetGroups(vpcId: string, subnetIds: string[]): Promise<AWS.
     }
 }
 
-export async function getDefaultAccountConfig(accountConfigParam: any): Promise<AccountConfig> {
-    const region = accountConfigParam.substring(accountConfigParam.indexOf('-') + 1);
-    const valid = await isValidRegion(region);
-    if (valid) {
-        // Set region so we can make calls to get the account VPC config (this will be set again later)
-        AWS.config.update({
-            region
-        });
-        const defaultVpc = await ec2Calls.getDefaultVpc();
-        const accountConfig: any = {
-            region,
-            vpc: defaultVpc.VpcId
-        };
+export async function getDefaultAccountConfig(region: string): Promise<AccountConfig> {
+    // Set region so we can make calls to get the account VPC config (this will be set again later)
+    process.env.AWS_REGION = region; // This will be set later on also
+    const defaultVpc = await ec2Calls.getDefaultVpc();
+    const accountConfig: any = {
+        region,
+        vpc: defaultVpc.VpcId
+    };
 
-        accountConfig.account_id = await stsCalls.getAccountId();
+    accountConfig.account_id = await stsCalls.getAccountId();
 
-        const subnets = await getDefaultVpcSubnets(accountConfig.vpc);
-        // The default VPC only has three public subnets, so we just have to use those for all the different tiers Handel supports
-        accountConfig.public_subnets = subnets;
-        accountConfig.private_subnets = subnets;
-        accountConfig.data_subnets = subnets;
+    const subnets = await getDefaultVpcSubnets(accountConfig.vpc);
+    // The default VPC only has three public subnets, so we just have to use those for all the different tiers Handel supports
+    accountConfig.public_subnets = subnets;
+    accountConfig.private_subnets = subnets;
+    accountConfig.data_subnets = subnets;
 
-        const cfStack = await getSubnetGroups(accountConfig.vpc, accountConfig.data_subnets);
-        accountConfig.rds_subnet_group = extensionSupport.awsCalls.cloudFormation.getOutput('RdsSubnetGroupName', cfStack);
-        accountConfig.elasticache_subnet_group = extensionSupport.awsCalls.cloudFormation.getOutput('ElastiCacheSubnetGroupName', cfStack);
-        return accountConfig as AccountConfig;
-    }
-    else {
-        throw new Error(`Invalid region: '${region}'`);
-    }
+    const cfStack = await getSubnetGroups(accountConfig.vpc, accountConfig.data_subnets);
+    accountConfig.rds_subnet_group = extensionSupport.awsCalls.cloudFormation.getOutput('RdsSubnetGroupName', cfStack);
+    accountConfig.elasticache_subnet_group = extensionSupport.awsCalls.cloudFormation.getOutput('ElastiCacheSubnetGroupName', cfStack);
+    return accountConfig as AccountConfig;
 }
